feat(core): Hot reload nodes on all servers
