+++ packages/cli/src/commands/base-command.ts
@@ -1,6 +1,8 @@
 import 'reflect-metadata';
 import { GlobalConfig } from '@n8n/config';
 import { Command, Errors } from '@oclif/core';
+import glob from 'fast-glob';
+import { access as fsAccess, realpath as fsRealPath } from 'fs/promises';
 import {
 	BinaryDataService,
 	InstanceSettings,
@@ -8,7 +10,13 @@ import {
 	DataDeduplicationService,
 	ErrorReporter,
 } from 'n8n-core';
-import { ApplicationError, ensureError, sleep } from 'n8n-workflow';
+import {
+	ApplicationError,
+	ensureError,
+	sleep,
+} from 'n8n-workflow';
+import path from 'path';
+import picocolors from 'picocolors';
 import { Container } from 'typedi';
 
 import type { AbstractServer } from '@/abstract-server';
@@ -40,6 +48,8 @@ export abstract class BaseCommand extends Command {
 
 	protected nodeTypes: NodeTypes;
 
+	protected loadNodesAndCredentials: LoadNodesAndCredentials;
+
 	protected instanceSettings: InstanceSettings = Container.get(InstanceSettings);
 
 	protected server?: AbstractServer;
@@ -67,7 +77,8 @@ export abstract class BaseCommand extends Command {
 		process.once('SIGINT', this.onTerminationSignal('SIGINT'));
 
 		this.nodeTypes = Container.get(NodeTypes);
-		await Container.get(LoadNodesAndCredentials).init();
+		this.loadNodesAndCredentials = Container.get(LoadNodesAndCredentials);
+		await this.loadNodesAndCredentials.init();
 
 		await Db.init().catch(
 			async (error: Error) => await this.exitWithCrash('There was an error initializing DB', error),
@@ -316,5 +327,59 @@ export abstract class BaseCommand extends Command {
 
 			clearTimeout(forceShutdownTimer);
 		};
+	}
+
+	protected async setupHotReload() {
+		if (!inDevelopment || process.env.N8N_DEV_RELOAD !== 'true') return;
+
+		const { default: debounce } = await import('lodash/debounce');
+		// eslint-disable-next-line import/no-extraneous-dependencies
+		const { watch } = await import('chokidar');
+
+		const { Push } = await import('@/push');
+		const push = Container.get(Push);
+
+		Object.values(this.loadNodesAndCredentials.loaders).forEach(async (loader) => {
+			try {
+				await fsAccess(loader.directory);
+			} catch {
+				// If directory doesn't exist, there is nothing to watch
+				return;
+			}
+
+			const realModulePath = path.join(await fsRealPath(loader.directory), path.sep);
+			const reloader = debounce(async (fileName: string) => {
+				console.info(
+					picocolors.green('⭮ Reloading'),
+					picocolors.bold(fileName),
+					'in',
+					loader.packageName,
+				);
+				const modulesToUnload = Object.keys(require.cache).filter((filePath) =>
+					filePath.startsWith(realModulePath),
+				);
+				modulesToUnload.forEach((filePath) => {
+					delete require.cache[filePath];
+				});
+
+				loader.reset();
+				await loader.loadAll();
+				await this.loadNodesAndCredentials.postProcessLoaders();
+				push.broadcast('nodeDescriptionUpdated', {});
+			}, 100);
+
+			const toWatch = loader.isLazyLoaded
+				? ['**/nodes.json', '**/credentials.json']
+				: ['**/*.js', '**/*.json'];
+			const files = await glob(toWatch, {
+				cwd: realModulePath,
+				ignore: ['node_modules/**'],
+			});
+			const watcher = watch(files, {
+				cwd: realModulePath,
+				ignoreInitial: true,
+			});
+			watcher.on('add', reloader).on('change', reloader).on('unlink', reloader);
+		});
 	}
 }
