+++ packages/cli/src/commands/start.ts
@@ -359,6 +359,8 @@ export class Start extends BaseCommand {
 				}
 			});
 		}
+
+		void this.setupHotReload();
 	}
 
 	async catch(error: Error) {
